Revert "fix: improve javascript file mime type detection / typing (#9971)"
This reverts commit ba416198

This is the most likely candidate for causing issue 10175:
A user's CSS/images are no longer working. I could not reproduce this

This is a temp revert to determine whether this commit is the problem

Reverts 9971
Fixes (Maybe) 10175
Reopens 9963
